requirements.txt
MetaTrader5==5.0.36
pandas==1.5.3
numpy==1.24.2
python-dotenv==1.0.0
pyyaml==6.0
logging
